fixed after review/simplify go-translator demo server
