+++ _common/ssl/acra-translator/acra-translator.crt
@@ -1,0 +1,26 @@
+-----BEGIN CERTIFICATE-----
+MIIEczCCA1ugAwIBAgIUdaqtl5Jh8nHPL+u+umX0UrO786QwDQYJKoZIhvcNAQEL
+BQAwdDELMAkGA1UEBhMCR0IxDzANBgNVBAgMBkxvbmRvbjEPMA0GA1UEBwwGTG9u
+ZG9uMRgwFgYDVQQKDA9HbG9iYWwgU2VjdXJpdHkxCzAJBgNVBAsMAklUMRwwGgYD
+VQQDDBNUZXN0IENBIGNlcnRpZmljYXRlMCAXDTIzMDgxNzA3NTkyMloYDzIwNzMw
+ODA0MDc1OTIyWjCBiDELMAkGA1UEBhMCR0IxDzANBgNVBAgMBkxvbmRvbjEPMA0G
+A1UEBwwGTG9uZG9uMRgwFgYDVQQKDA9HbG9iYWwgU2VjdXJpdHkxCzAJBgNVBAsM
+AklUMTAwLgYDVQQDDCdUZXN0IGxlYWYgY2VydGlmaWNhdGUgKGFjcmEtdHJhbnNs
+YXRvcikwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQC8M66fdDRtAQd3
+s43POrRrsjzohJ1Z4GsOljqdU1K4/K3nqscJawNcz8lkafE0GVCwrEpGgJA3Dbrh
+K8gGI8CuU22GDwzobd4jerI3bsB2k4IQ6tooya8G6SELd3nJjDZP1VEmjdO1+BZD
+9VK1tfbMckGw2qLoFjBHc+h14k2eUqK+gBeR8n8URF44DF+gYrgDWh+3X9X64Wj8
+Pu8AQJIDYxMvDE4vriyKxpqWJUjj1gDTHCZmaJATA5hVDuclb3uwFiY+iFwUw3rG
+I+tBOpc8N72qB1+hylupPSQEB9u4PWLk8YHJiQ1Y/QKpbKaO7KhYIQr9KTvz7SEV
+DmXJs9NNAgMBAAGjgeUwgeIwCQYDVR0TBAIwADALBgNVHQ8EBAMCBeAwJQYDVR0R
+BB4wHIIJbG9jYWxob3N0gg9hY3JhLXRyYW5zbGF0b3IwMQYIKwYBBQUHAQEEJTAj
+MCEGCCsGAQUFBzABhhVodHRwOi8vMTI3LjAuMC4xOjg4ODgwLgYDVR0fBCcwJTAj
+oCGgH4YdaHR0cDovLzEyNy4wLjAuMTo4ODg5L2NybC5wZW0wHQYDVR0OBBYEFF3M
+Qjc/RBv9TcpwvKNKw5tlVy+sMB8GA1UdIwQYMBaAFOlYUMjmJWHi1rJRWyG9SbuD
+oEqoMA0GCSqGSIb3DQEBCwUAA4IBAQAp+IjJcwrpMrNi5w6zP3CbchT4e4NEEe/6
+P6tYM9hA1/xlI9+9vsvr1/1Eq5ss5TpPazt/D9a2v8nxvc5O9qzA4zdDWYs0cs7Z
+f5wxgmLkEdMWkTh2ofGLEAzDUSAQjZHE5IH0ukkfbrT37WkINkyQcbIP3abr7/X6
+cmJWMMU+csDc1za/PirU34DeEHDzOkbcwpv9K/H6MLiNe6ZBReuLQRtYg0tUr5E0
+tH8JsINcnpTwrYcdAEQUDRM8l2piqvuXr5qzAMRrDfy19vGHtosf4ZG6AWVIFOVX
+W48wEmYtoeUrIpsN+eTPbsPfWn36TKYPBAzCrKvZ+6jxeOsx/t+p
+-----END CERTIFICATE-----
