+++ _common/ssl/acra-translator/acra-translator.key
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQC8M66fdDRtAQd3
+s43POrRrsjzohJ1Z4GsOljqdU1K4/K3nqscJawNcz8lkafE0GVCwrEpGgJA3Dbrh
+K8gGI8CuU22GDwzobd4jerI3bsB2k4IQ6tooya8G6SELd3nJjDZP1VEmjdO1+BZD
+9VK1tfbMckGw2qLoFjBHc+h14k2eUqK+gBeR8n8URF44DF+gYrgDWh+3X9X64Wj8
+Pu8AQJIDYxMvDE4vriyKxpqWJUjj1gDTHCZmaJATA5hVDuclb3uwFiY+iFwUw3rG
+I+tBOpc8N72qB1+hylupPSQEB9u4PWLk8YHJiQ1Y/QKpbKaO7KhYIQr9KTvz7SEV
+DmXJs9NNAgMBAAECggEAElPddWAde4tbnPpqh3FEllj7ryW8UGnXPcrpzLNgvWy3
+SiBqlKNhZvLsk6+bfS2FRZyUTncA/v4tIkt0sGBf8w9jKD3OoyiRNuxLxF6FNfxE
+cNL7dCMRPT+LvRr2obwv99NNZ/ns0ZjmINptomN6NCELqPuJshyjDqTze4i9HVCX
+pXiXWAx9qpbxbO/JD0wePbXKzrnEAfEoXs4Izy/nSOZ5CznbC11ycJZuRPsOB5Vx
+81qcVMx3z5BdkV9/yY0Y0/1vjdLBrHCvKNjMji+VCFqF5efQJUAlwgJ/SHzppERX
+tKVxGeMUWgj74bywu23upkKpoonNRN8soXp1F5Fr0wKBgQDNq1/TdoM6t0BQ0Rox
+l+M6AaaQdPkW8zWm1/uvA9E0lyl4FYmmJe9ToNFjg1j6krLa1j12kjLqm7ccG964
+z0CXl1y1tcedTWYx8x162vV+9CjEO1RhV+fX4hI/4CxCNTCF3d7Rz4JBNWQDIHdn
+P2a/ddAl2u58g+yVDri4USz9hwKBgQDqQgOL9+3RVvDx9rz945n25aRSt3U7BvSm
+I3G4tq98FuIyp2xUfWGiqU2KUMqKZGKhbSrAWhjqTZ9MIhgxxVgeDEQJkEZYI2Lx
+YtdDzXBEMyNSK1D6aJlDfhm0dFdFvOs18Uea3xIlaqRmYtXmtKNlN1O6JVjtH+GS
+6vfd4ns9iwKBgF5kLsCgfa3u+KeesETejrzKkSVvafqseI0wRySVvehDEBc620fT
+1Tkub6saVzX5KfjhpqfKhW/DM5DFAb0T8XMhVr1Wxlxdb2+xQnBmGZHAgQ0+pYVf
+kNo1uWZzfhyMyM5He8ITWJaQ6Gu+9M3//SoV6gfVqOqgqW3nqcN+DgX1AoGBAJic
+TwE1kLoLYMbRTog8aYZeTrIQkI/W1TUS7DhA04wMwGnthnZ4hE3a7AgK4BhOk67E
+ltBfIdIqpXeM0qRCA0j4Bq8QTbOHwiKtmL20LcpQYldantOzpv1APnDARkbJb5Zh
+zJtJbeHwWo/VHbLBxZzf8WzYfdllJqh26OmlbVJHAoGBAKLNx17wfnVfT80at2Lp
+1r2jvZREYkkA79T6PniE+1C/Mo8L34NML05SHu5nC3DBl4WSv4Jr8qHIG/6+56J7
+IwcZrN0/u+ibIpCuJBoYPjr07rKzH/vIOdbIvT3+aFK6DmYoQ+DybHNOkxCVOKEk
+XEQJwR/m7+gkno0aAjZf2PDd
+-----END PRIVATE KEY-----
