+++ _common/ssl/acra-translator/root.crt
@@ -1,0 +1,23 @@
+-----BEGIN CERTIFICATE-----
+MIID4TCCAsmgAwIBAgIUQ24MW+Sq36AfAUy6Rq1hRYHI7CgwDQYJKoZIhvcNAQEL
+BQAwdDELMAkGA1UEBhMCR0IxDzANBgNVBAgMBkxvbmRvbjEPMA0GA1UEBwwGTG9u
+ZG9uMRgwFgYDVQQKDA9HbG9iYWwgU2VjdXJpdHkxCzAJBgNVBAsMAklUMRwwGgYD
+VQQDDBNUZXN0IENBIGNlcnRpZmljYXRlMCAXDTIwMDkyMjE2MDUzNFoYDzIwNzAw
+OTEwMTYwNTM0WjB0MQswCQYDVQQGEwJHQjEPMA0GA1UECAwGTG9uZG9uMQ8wDQYD
+VQQHDAZMb25kb24xGDAWBgNVBAoMD0dsb2JhbCBTZWN1cml0eTELMAkGA1UECwwC
+SVQxHDAaBgNVBAMME1Rlc3QgQ0EgY2VydGlmaWNhdGUwggEiMA0GCSqGSIb3DQEB
+AQUAA4IBDwAwggEKAoIBAQC+EmCoONyAvjAvWwqxVTaUCkPDAzo2f67xg+hppOJz
+JuBOXed5zBvfStGITIWvCzStrSK6PSLE6J6ZM1FNjplJaD/uRORRWA83SAqqOszj
+yDz+Qds7p/zsqEu9maBWO3MK0FLl9cENh4cH+XhM/qs8CXsGpDz6HMDUgydbRCdk
+dlUBs+kwK4amjhKe3mWbJSKdidqYA/NHzYTuWISYydKi4zGpRxuo3Ajrc/OuztR4
+ZroZjpFOQILnY2q9FwVWRR+3lsZ8EJVGB2TTCMnOVnB03HdGVsQpxgon3fMeQyd6
+vDaHnIw48XUhBPaVoasr7iVInR9xkZq/HloXSHqZhpgLAgMBAAGjaTBnMA8GA1Ud
+EwEB/wQFMAMBAf8wFAYDVR0RBA0wC4IJbG9jYWxob3N0MB0GA1UdDgQWBBTpWFDI
+5iVh4tayUVshvUm7g6BKqDAfBgNVHSMEGDAWgBTpWFDI5iVh4tayUVshvUm7g6BK
+qDANBgkqhkiG9w0BAQsFAAOCAQEABqiGW6PA4MEECFmKu82YrAQrVN20696pnIbJ
+PTLNoLdagj4tXBcRXO87g9enfuJfXAL26FJHvfnwzDIkHkIi10oBJZLjQ/6BQwR+
+jDSEXeRfDajA6LhO5iDRp9C/DvfzcnpW6yIOE+ugdrroCHGk5k5abjwGISuobJgT
+czDf+qWpAGEp1JoVPt4QaCKnFwNbXJa37XfyQajP7EVZlUHKr4FWcaNd2ZhzTVMt
+QlMzv5Dqc7CBT7a1+aLtuHEX1BdZRBdNf9Owv7DvbiJO3dMq29tr+RGyj8jFLvi7
+CUW9vw/S1rYOsG+tgdQc6Wpzj3kupRdfwDWT25NYMyWdV5wQ6A==
+-----END CERTIFICATE-----
