--- acra-translator/go-translator-demo/Makefile
@@ -1,5 +1,0 @@
-swagger:
-	cd ./cmd/api/ && swag init --parseDependency --output swagger --overridesFile swagger/.swaggo
-
-swagger-format:
-	cd ./cmd/api/ && swag fmt
