--- acra-translator/go-translator-demo/cmd/api/.env.example
@@ -1,10 +1,0 @@
-ADDRESS=0.0.0.0:8008
-# overwrite swagger host, use `ADDRESS` if not set
-SWAGGER_HOST=0.0.0.0:8008
-
-# debug, info, warn, error are acceptable
-LOG_LEVEL=info
-
-MONGODB_CONNECTION_URL="mongodb://root:password@localhost:27017/demo?authSource=admin"
-
-ACRA_TRANSLATOR_URL=
