+++ acra-translator/go-translator-demo/repositories/init.go
@@ -1,4 +1,4 @@
-package persistence
+package repositories
 
 import (
 	"context"
